requirements.txt
## python=3.9.0; cuda=11.8
# torch==2.0.1 # pip install torch==2.0.1+cu118 torchvision==0.15.2+cu118 torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.15.2
# torchaudio==2.0.2
byted-cruise==0.5.0
opencv_python
Pillow==9.3.0
pytorch_metric_learning
jieba
jinja2
timm
transformers==4.33.2
fairscale
pycocoevalcap
ruamel_yaml
byted_ocr_datakit
einops
jsonargparse[signatures]
fsspec
opencv-python-headless
sentencepiece
pyarrow
pandas
seqeval
nltk
rouge
deepspeed==0.14.2
byted-wandb==0.13.67
accelerate==0.31.0 # baichuan2
bitsandbytes
xformers==0.0.21 # baichuan2
peft
ftfy # for clip
apex
openpyxl==3.1.4 # for mm bench
